refactor: RandomLadderStepGenerator 패키지 이동
