+++ src/main/java/ladder/domain/generator/RandomLadderStepGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.generator;
 
 import ladder.domain.generator.LadderStepGenerator;
