+++ src/test/java/ladder/domain/generator/RandomLadderStepGeneratorTest.java
@@ -1,8 +1,6 @@
 package ladder.domain.generator;
 
-import ladder.domain.RandomLadderStepGenerator;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
